requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
fastapi==0.115.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.3.7
langchain-core==0.3.63
langchain-openai==0.2.9
langchain-text-splitters==0.3.8
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langsmith==0.1.147
msgpack==1.1.2
multidict==6.7.1
numpy==1.26.4
openai==1.54.0
orjson==3.11.8
packaging==24.2
pluggy==1.6.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.41.3
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.0
yarl==1.23.0
